Tabs working, components rendering
Moved components to their respective folders.

Reworked the routing and the NavBar a bit. Wrapped the react-routing
components in a `LinkContainer` according to the
[react-bootstrap](https://github.com/react-bootstrap/react-router-bootstrap)
docs.

Deleted a few needless components. Moved the Routes from their own Index
into the Index I'm rendering from.
