+++ app/assets/stylesheets/application.css
@@ -13,5 +13,5 @@
  *= require_tree .
  *= require_self
  */
- @import "~bootstrap/scss/bootstrap";
- @import '~bootstrap/dist/css/bootstrap.min.css';
+ /* @import "~bootstrap/scss/bootstrap";
+ @import '~bootstrap/dist/css/bootstrap.min.css'; */
