--- app/controllers/landlords_controller.rb
@@ -1,9 +1,0 @@
-class LandlordsController < ApplicationController
-    def index
-      @landlords = Landlord.all
-    end
-
-    def show
-        @landlord = Landlord.find(params[:id])
-    end
-end
