+++ app/javascript/components/App.tsx
@@ -1,0 +1,14 @@
+import * as React from "react";
+import NavBar from "./NavBar";
+import Home from "./Home";
+import { Outlet } from "react-router";
+
+export default function App() {
+  return (
+    <>
+      <Home />
+      <NavBar />
+      <Outlet />
+    </>
+  );
+}
