+++ app/javascript/components/Landlord/Landlord.tsx
@@ -1,0 +1,5 @@
+import * as React from "react";
+
+export default function Landlord() {
+  return <div>Landlord</div>;
+}
