+++ app/javascript/components/Landlord/Landlords.tsx
@@ -1,0 +1,7 @@
+import * as React from "react";
+
+export default function Landlords() {
+  return (
+    (<div>Landlords</div>)
+  );
+}
