+++ app/javascript/components/NotFound.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react'
+
+export default function NotFound() {
+  return (
+    <div>NotFound</div>
+  )
+}
