+++ app/javascript/components/Property/Property.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react'
+
+export default function Property() {
+  return (
+    <div>Property</div>
+  )
+}
