+++ app/javascript/components/Property/PropertyLoader.tsx
