--- app/javascript/routes/Landlords.tsx
@@ -1,7 +1,0 @@
-import * as React from 'react'
-
-export default function Landlords() {
-  return (
-    <div>Landlords</div>
-  )
-}
