+++ app/views/layouts/application.html.erb
@@ -8,6 +8,7 @@
     <%= javascript_include_tag "application", "data-turbo-track": "reload", defer: true %>
     <%= stylesheet_link_tag 'application', media: 'all', 'data-turbolinks-track': 'reload' %>
     <%= stylesheet_link_tag "https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" %>
+    <%= stylesheet_link_tag "https://unpkg.com/react-quill@1.3.3/dist/quill.snow.css" %>
   </head>
 
   <body>
