+++ package.json
@@ -19,6 +19,7 @@
     "@types/react-bootstrap": "^0.32.29",
     "@types/react-dom": "^17.0.11",
     "@types/react-router": "^5.1.17",
+    "@types/react-router-bootstrap": "^0.24.5",
     "@types/react-router-dom": "^5.3.2",
     "@types/react-table": "^7.7.9",
     "@types/styled-components": "^5.1.23",
@@ -40,8 +41,10 @@
     "react-bootstrap": "^2.1.2",
     "react-dom": "^17.0.2",
     "react-is": "^17.0.2",
+    "react-quill": "^1.3.5",
     "react-redux": "^7.2.6",
     "react-router": "^6.0.2",
+    "react-router-bootstrap": "^0.26.0",
     "react-router-dom": "^6.0.2",
     "react-scripts": "^5.0.0",
     "react-table": "^7.7.0",
