+++ tsconfig.json
@@ -7,7 +7,7 @@
     "module": "es6",
     "moduleResolution": "node",
     "sourceMap": true,
-    "target": "es5",
+    "target": "es6",
     "jsx": "react",
     "noEmit": true,
     "allowJs": true,
